处理ArticleTag实体类和 Tag的关系

+++ src/main/java/com/akbar/service/impl/TagServiceImpl.java
@@ -1,6 +1,10 @@
 package com.akbar.service.impl;
 
 import com.akbar.annotation.RequiresAdmin;
+import com.akbar.constant.MessageConstant;
+import com.akbar.exception.DeletionNotAllowedException;
+import com.akbar.mapper.ArticleMapper;
+import com.akbar.mapper.ArticleTagMapper;
 import com.akbar.pojo.entity.Tag;
 import com.akbar.mapper.TagMapper;
 import com.akbar.service.TagService;
@@ -14,6 +18,12 @@ public class TagServiceImpl implements TagService {
 
     @Autowired
     private TagMapper tagMapper;
+
+    @Autowired
+    private ArticleMapper articleMapper;
+
+    @Autowired
+    private ArticleTagMapper articleTagMapper;
 
     /**
      * 添加标签
@@ -54,6 +64,15 @@ public class TagServiceImpl implements TagService {
     @RequiresAdmin
     @Override
     public void deleteTag(Integer id) {
+        // 当前标签下有文章，不能删
+        Integer count = articleMapper.countByTagId(id);
+        if (count > 0) {
+            throw new DeletionNotAllowedException(MessageConstant.TAG_HAS_ASSOCIATED_ARTICLES);
+        }
+
+        // 先删除 article_tag 中的关联记录
+        articleTagMapper.deleteByTagId(id);
+        // 再删除 tag 表中的记录
         tagMapper.delete(id);
     }
